requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-googlemaps @ file:///Users/xjxc391/Documents/personal/Flask-GoogleMaps
future==0.18.2
htmlmin==0.1.12
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsmin==2.2.2
livereload==2.6.3
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==5.5.12
mkdocs-material-extensions==1.0
mkdocs-minify-plugin==0.3.0
nltk==3.5
numpy==1.19.1
pandas==1.1.1
Pygments==2.6.1
pymdown-extensions==8.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
tornado==6.0.4
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
